# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API",
        "description": "Documentation API complète pour la plateforme LMS EduFlow. Gestion des cours, étudiants, enseignants, favoris, inscriptions et groupes.",
        "contact": {
            "name": "EduFlow Support",
            "email": "support@eduflow.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{host}:{port}/api",
            "description": "EduFlow API Server",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "enum": [
                        "localhost",
                        "api.eduflow.dev"
                    ],
                    "default": "localhost"
                },
                "port": {
                    "enum": [
                        "8000",
                        "443"
                    ],
                    "default": "8000"
                }
            }
        }
    ],
    "paths": {
...
        "/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Lister les cours favoris",
                "description": "Récupère la liste paginée des cours favoris de l'étudiant connecté",
                "operationId": "getFavoritesList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de résultats par page (défaut: 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des cours favoris récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "format": "int64",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "PHP Avancé"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Cours approfondi de PHP avec patterns modernes"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 49.99
                                            },
                                            "teacher_id": {
                                                "type": "integer",
                                                "format": "int64",
                                                "example": 2
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "format": "int64",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé - Rôle student requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/favorites/{course}": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Ajouter un cours aux favoris",
                "description": "Ajoute un cours à la liste des favoris de l'étudiant connecté",
                "operationId": "addFavorite",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID du cours à ajouter aux favoris",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cours ajouté aux favoris avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course added to favorites successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cours non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflit - Cours déjà dans les favoris",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "This course is already in your favorites."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé - Rôle student requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Supprimer un cours des favoris",
                "description": "Supprime un cours de la liste des favoris de l'étudiant connecté",
                "operationId": "removeFavorite",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID du cours à supprimer des favoris",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cours supprimé des favoris avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course removed from favorites successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cours non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflit - Cours n'est pas dans les favoris",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "This course is not in your favorites."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé - Rôle student requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token pour l'authentification JWT",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}